75000 edition on windows

+++ PrimeC/solution_5/sieve_helpers.h
@@ -11,7 +11,7 @@
 #endif
 
 #define bitshift_t uint64_t // type used to shift bits
-#define counter_t  int32_t  // type used to count loops, etc. Some processors/compilers are faster at 32 bits
+#define counter_t  uint64_t  // type used to count loops, etc. Some processors/compilers are faster at 32 bits
 
 #if defined(counter_t) && (counter_t == int32_t || counter_t == uint32_t)
     #define COUNTER_T_MAX_SAFE_VALUE 1000000000ULL
